Collections: deprecate $OONameTokens in favour of namespacedNameTokens()
Deprecate the `$OONameTokens` property in favour of a new `Collections::namespacedNameTokens()` method.

PHP 8.0 introduces three new tokens to represent identifier names. As those tokens will not always be available, a property can't handle this.

In anticipation of the introduction of the new tokens, I'm deprecating the `Collections::$OONameTokens` property to prevent BC breaks at a later point in time.

Includes adding unit tests for the new method.

Includes switching out existing uses of the deprecated property for the new method.