Collections::propertyTypeTokens[BC](): support the PHP 8 identifier name tokens
Includes adjusted unit test.

This commit implicitly adds support for PHP 8 identifier name tokens to the `Variables::getMemberProperties()` method.

The existing unit tests already cover this.

This commit also _silently_ adds support for PHP 8 identifier name tokens to the `BCFile::getMemberProperties()` method. The PHP 8 identifier name tokens are not supported yet in PHPCS itself and until they are, the fact that the `BCFile::getMemberProperties()` method supports them should be regarded as an artefact and not as official support.