BCFile/GetTokensAsString/GetTokensAsStringTest: make tests independent of PHP 8 identifier names
Related to 201

As the token counts for identifier names are different between PHP < 8 and PHP 8 and depending on which PHPCS version will be used, the test for the `BCFile::getTokensAsString()` method is now unstable.

This commit adds a new test case (calculation) which doesn't suffer from that problem to allow for testing with exact token lengths, the same as was previously done with the namespace declaration test.

Includes:
* In `BCFile/GetTokensAsStringTest`: add an additional test to still cover the namespace declaration test case.
* In `GetTokensAsString/GetTokensAsStringTest`: add a test for the new calculation test case.