Tests: skip select tests when the PHP 8 identifier name tokenization is used
When the PHP 8 identifier name tokenization is used, the `T_NAMESPACE` token for the `namespace` keyword when used as an operator will no longer exist. Similarly, the `T_STRING` for `name` in a `namespace\name` snippet will no longer exist.

For certain test cases were setup specifically to prevent false positives/negatives for that type of code.

As the "target" token which would be passed to the functions which could have these false positives/negatives, won't exist when the PHP 8 identifier name tokenization is used, these false positives/negatives won't occur and the test becomes redundant.

This commit adds selective test skipping for a limited number of test cases where this applies.