PassedParameters::hasParameters(): compatibility with the PHP 8 identifier name tokenization
The `PassedParameters::hasParameters()` does a token based check on whether or not the passed token is one which is handled by this method, as well as the other methods in this class.

That check now needs to take the PHP 8 identifier name tokens into account and should treat them as the same as `T_STRING`, i.e. as as suspected function call.

This commit adjusts the `PassedParameters::hasParameters()` method to allow for these new tokens.

Includes additional unit tests covering the different tokens.