Operators::isTypeUnion(): BC-fix - namespace operator in return type
The `Operators::isTypeUnion()` method is subseptible to the Tokenizer bug which will be fixed in upstream PR 3066 / PHPCS 3.5.7, where the scope opener/closer/condition would not be set for a function declared with a namespace relative return type using the `namespace` keyword as an operator.

A work-around has been added to the method now.

Includes unit test.

Includes minor fix to existing inline documentation.