BCFile::getMethodParameters(): sync with PHPCS / namespace operator in types
PR 180 added support for the namespace keyword used as an operator in type declarations to the `FunctionDeclarations::getParameters()` method.

Upstream PR 3066 added the same to the PHPCS native `File::getMethodParameters()` method.

As that PR has now been merged, this commit syncs the upstream changes into the `BCFile::getMethodParameters()` method, moves the unit tests to the BCFile test class and updates the documentation for the `FunctionDeclarations::getParameters()` method to support the changes as have been merged upstream.