BCFile::getMethodProperties(): sync with PHPCS / namespace operator in types
PR 180 added support for the namespace keyword used as an operator in type declarations to the `FunctionDeclarations::getProperties()` method and by extension the `BCFile::getMethodProperties()` method, even though PHPCS upstream did not support this yet.

Upstream PR 3066 added the same to the PHPCS native `File::getMethodProperties()` method.

As that PR has now been merged, this commit syncs the upstream changes in, by moving the unit tests from the "Diff" tests to the BCFile tests and annotating that namespace operators in type declarations are now officially supported in the `BCFile::getMethodProperties()` method.