Tests: merge the `isShortArray()` and `isShortList()` tests
A square open/close bracket is either a short array or a short list or neither (= plain square bracket) but it can never be both.

With that in mind, there is (and was) a significant overlap being the test cases which needed to be covered for the `isShortArray()` and the `isShortList()` methods.

This was in part already acknowledged by the `IsShortArrayTokenizerBC1Test` and the `IsShortListTokenizerBC1Test` classes and the `IsShortArrayTokenizerBC2Test` and the `IsShortListOrArrayTokenizerBC2Test` classes already using the same test case files.

This change iterates on this further, by making sure that the `isShortArray()` and `isShortList()` methods can never both return `true` for any one test case.

To that end, the test files for the `isShortArray()`and the `isShortList()` methods are not completely merged together.

All existing test cases are now tested against both methods (most already were) and a safety mechanism has been put in place to ensure that the methods can never return `true` for the same test case for both methods.