Arrays::isShortArray()/Lists::isShortList(): work round PHP8 interpolated string dereferencing
As of PHP 8, text strings containing interpolated variables can be dereferenced, however, the square brackets in `"foo{$bar}"[0]` are incorrectly tokenized as _short array_ brackets instead of as "normal" square brackets.

A fix for the PHPCS Tokenizer has been pulled upstream.

Fingers crossed, the fix will be merged soon.

In the mean time, this commit fixes the issue for PHPCS 2.6.0 - current.

Once the upstream PR has been merged, the version numbers in the inline documentation and test marker will need to be updated.

Includes unit test.

Refs:
* https://wiki.php.net/rfc/variable_syntax_tweaks#interpolated_and_non-interpolated_strings
* squizlabs/PHP_CodeSniffer 3172