PassedParameters::getParameters(): fix backward compatibility with PHPCS < 3.6.0
To support named parameters in older PHPCS versions in which named parameters aren't supported by PHPCS itself yet, some fixes and tweaks are needed.

Depending on the specific code being used, in _most_ cases, a parameter name label in older PHPCS versions would be tokenized as `T_GOTO_LABEL` with the `T_COLON` token being removed and the `:` being added to the "goto label" content.

This adds a BC support to the `getParameters()` method to allow for tokenization of the parameter name as the `T_GOTO_LABEL` token, as well as for the `T_COLON` being tokenized separately, but as the wrong token - `T_STRING`/`T_INLINE_ELSE` -.

As the `T_GOTO_LABEL` token swallows the `T_COLON` token, this also means that the token positions returned by `PassedParameters::getParameter()` will be different depending on whether sniffs are run on a PHPCS version supporting named parameters or not.

The unit tests have been adjusted to take this into account.