PassedParameters::getParameters(): fix compatibility with PHP 8.0 (tests only)
While the PHP 8.0 `match` control structure is not yet supported in PHPCS itself, in PHP 8.0, it will be recognized as a keyword, which will - again - throw off the token positions in the tests.

The `PassedParameters::getParameters()` method itself already handles "`match` as a keyword vs not a keyword" correctly.

This adds a tweak to the unit tests to work-around the different expected token positions.
Once PHPCS adds support for `match` control structures, this tweak will most likely need adjusting.