Tests/UtilityMethodTestCase::skipJSCSSTestsOnPHPCS4: get rid of "risky test" warning
When the `@doesNotPerformAssertions` annotation is used, the test shows as risky when run in combination with PHPCS 4.x as it does perform assertions (expects an exception), but removing the annotation would make the test as risky when run in combination with PHPCS 3.x as in that case, it didn't perform an assertion.

While a little hackish, this small tweak gets rid of the warnings in all circumstances.