ObjectDeclarations::getClassProperties(): allow for `abstract final`
Having both `abstract` as well as `final` as a modifier for a class is an oxymoron (and a parse error in PHP), but the utility methods should not have an opinion on that.

The original `BCFile::getClassProperties()` method already handled this correctly.

Now the `ObjectDeclarations::getClassProperties()` method does as well.

Includes moving a test from the "diff" test files to the generic test files.