PassedParameters::getParameters(): add new, optional $limit parameter
This adds a new, optional `$limit` parameter to the `PassedParameters::getParameters()` method to allow for limiting the number of parameters to be parsed.

This allows for higher efficiency when retrieving the parameters/array entries, especially for large arrays if only the first few entries need to be examined.

Includes implementing the use of this parameter in the `PassedParameters::getParameter()` method.

Includes dedicated unit tests for `PassedParameters::getParameters()` concerning the use of the `$limit` parameter.

The change in the `PassedParameters::getParameter()` method is already largely covered by the existing unit tests.
Only some extra tests for the use of `$limit` in combination with "named parameters" were needed. Those have been added to the `GetParameterFromStackTest` class.

Includes some minor documentation and test function naming improvements to the `GetParameterFromStackTest` class.