PassedParameters::[has|get]Parameters(): add new, optional `$isShortArray` parameter
This adds a new, optional `$isShortArray` parameter to the `PassedParameters::hasParameters()` and the `PassedParameters::getParameters()` methods to allow for skipping the "is short array" check for passed square brackets when the value is set to `true`.

It is HIGHLY DISCOURAGED to ever use to this parameter and this parameter should only ever be used with extreme care and forethought!

The reason to add this parameter is that it will allow for the "is short array or short list" check to use the `PassedParameters` functionality without throwing execution into an infinite loop.
Most likely, the use in that check will be the only legitimate use of the parameter.

Includes dedicated unit tests for documenting the behaviour of the class when the `$isShortArray` parameter is used.