UseStatements::splitImportUseStatement: improve test consistency
The "parse error due to reserved keyword as alias" tests for `use .... as function` and `use ... as class` already showed what the _real_ expected output should be.

The test for `use ... as const` did not, as the code after it did not contain a semi-colon.

As part of the underlying issue for this code sample (the semi-colon being tokenized as `T_STRING`) has been fixed in PHPCS 3.7.0, this now meant that the test would give a different result in PHPCS 3.7.0+ vs PHPCS < 3.7.0.

As use statements are generally at the top of a file, it would be extraordinarily rare for code like that to ever occur in real life. On top of that, this is a parse error.
With both of those things in mind, just fixing the test code to allow the test to return consistent results PHPCS cross-version feels like the appropriate fix.