Update for Composer 2.2
The `dealerdirect/phpcodesniffer-composer-installer` Composer plugin is a non-dev requirement for PHPCSUtils.

As of Composer 2.2.0, Composer plugins need to be explicitly allowed to run.

This commit updates the installation instructions to handle this.
It also adds an info block suggesting how to handle this for external standards including PHPCSUtils.

Includes updating the URL references to the plugin to point to the new home of the plugin. (see [upstream issue #146](https://github.com/PHPCSStandards/composer-installer/issues/146))

Refs:
* https://blog.packagist.com/composer-2-2/#more-secure-plugin-execution