GH Actions: pass the PHPCS version to test runs
Only just realized that the test for the `BackCompat\Helper::getVersion()` method wasn't running since the move to GH Actions as it expects the environment variable `PHPCS_VERSION` to be set, which was set automatically from the matrix in Travis, but needs to be explicitly passed in GH Actions.

Fixed now.