GH Actions/coverage: add extra (debug) step
While the feedback shown in the `setup-php` step is awesome, it doesn't show the version of Xdebug used when running code coverage, while, in case of issues, the version number of Xdebug is crucial to know.

This extra step allows for that information to always be available.

I've only added the step to the code coverage job though as that's the only one for which the Xdebug version is relevant.