Collections: remove $arrayOpenTokensBC in favour of arrayOpenTokensBC()
Over the last three PHP versions, PHP has steadily kept on introducing new tokens.

In the context of this class, the introduction of a new token means that a collection can no longer be a property and must be changed to a method with some logic instead.

With this in mind, all properties in the `Collections` class are being deprecated in favour of methods, to prevent the BC-breaks this kind of change would cause if left till later.

The deprecated properties will be removed (or made `private`) before the `1.0.0` version is tagged. The deprecation in the upcoming alpha gives standards which have started to implement the use of PHPCSUtils a little time to switch over though.

**_Note: as this property has been introduced during the `alpha4` dev cycle, this property is being outright removed instead of being deprecated._**

Includes only a perfunctory unit test for the new method, as the `arrayOpenTokensBC()` method doesn't contain any actual logic.
For now, it is just safeguarded that the method is available, can be called and returns the expected type of return value.

Includes replacing any references to this property in the PHPCSUtils codebase to now call the function instead.