Collections: deprecate $magicConstants in favour of BCTokens::magicConstants()
Over the last three PHP versions, PHP has steadily kept on introducing new tokens.

In the context of this class, the introduction of a new token means that a collection can no longer be a property and must be changed to a method with some logic instead.

With this in mind, all properties in the `Collections` class are being deprecated in favour of methods, to prevent the BC-breaks this kind of change would cause if left till later.

The deprecated properties will be removed (or made `private`) before the `1.0.0` version is tagged. The deprecation in the upcoming alpha gives standards which have started to implement the use of PHPCSUtils a little time to switch over though.

**Note**: As PR 172 already introduced a method for the same in the `BCTokens` class as PHPCS 3.5.6 introduced a `Tokens` property for the same, no replacement has been added to the `Collections` class.

Includes replacing any references to this property in the PHPCSUtils codebase to now call the function in `BCTokens` instead.