:sparkles: New (internal) `Cache` and `NoFileCache` classes
... to allow for caching the results of processing intensive utility methods.

The `Cache` class is intended to be used for utility functions which depend on the `$phpcsFile` object.
The `NoFileCache` class is for file-independent functions.

Use selectively and with care as the memory usage of PHPCS will increase when using these caches!

Includes full set of tests to cover this new functionality.

These new tests have been added to the `RunFirst` test suite to prevent the cache setting and clearing done in these tests from interfering with the rest of the tests.