PassedParameters::hasParameters(): account for upstream tokenization change for `parent`
Upstream PR squizlabs/PHP_CodeSniffer 3546, which is included in PHPCS 3.7.0, changed the tokenization of the `parent` keyword in `new parent()` from `T_STRING` to `T_PARENT`.

This has consequences for the `PassedParameters::hasParameters()` method and associated methods:
* The `Collections::parameterPassingTokens()` method will now need to include the `T_PARENT` token.
* The underlying `Collections::functionCallTokens()` method should as well.
* The `PassedParameters::hasParameters()` method now needs to allow for the `T_PARENT` token potentially being used as part of a `new parent()` function call.

This commit fixes all that up.

Includes additional unit tests to safeguard it all (and some more).