BCTokens::functionNameTokens(): sync with upstream
Upstream PR squizlabs/PHP_CodeSniffer 3546, which is included in PHPCS 3.7.0, changed the tokenization of the `parent` keyword in `new parent()` from `T_STRING` to `T_PARENT`.

Follow-up PR squizlabs/PHP_CodeSniffer 3619, which is expected to be included in PHPCS 3.7.2, adds the `T_PARENT` token to the `Tokens::$functionNameTokens` array.

This syncs the `BCTokens::functionNameTokens()` function with this upstream change.