Tests/bootstrap: allow for running the tests with/without caching
This commit sets up a mechanism in the test `bootstrap.php` file to turn the cache on/off depending on an environment variable, which can be set from within a `phpunit.xml` file or on the OS-level.

This allows for running the tests both with caching turned on, as well as with caching turned off.

By default, the tests will run with caching turned **off**.

Being able to run the tests with both settings will allow for:
* [Caching off] Making sure that all code paths can be reached (code coverage check). Caching being enabled could prevent some code paths from being tested.
* [Caching on] Making sure that caching does not (negatively) impact the actual results of the functions.