ControlStructures::getDeclareScopeOpenClose(): implement use of the new `Cache` class
As `declare` constructs using alternative syntax can be long, determining the scope opener/closer can involve a lot of token walking, which is why I'm implementing caching for it.

Includes some minor tweaks to the exit routes of the function to ensure that the cache will always be set when the `T_DECLARE` token doesn't natively have a scope opener/closer assigned.

Includes a dedicated test to verify the cache is used and working.