Arrays::getDoubleArrowPtr(): implement use of the new `Cache` class
While this function will generally be _fast_ for array entries _with_ a double arrow, it will be _slow_ for array entries without a double arrow, especially if the contents of the array item contains a lot of tokens as the fact that the array item does not have a double arrow can only be determined when the end of the array item has been reached.
With this in mind, I'm implementing caching for it.

Includes some minor tweaks to the exit routes of the function to ensure that the cache will always be set.

Includes a dedicated test to verify the cache is used and working.