Arrays::isShortArray()/Lists::isShortList(): implement use of the new `Cache` class
These are easily the most used functions as well as the slowest functions (when dealing with large arrays/lists).

While this commit already implements the use of the new `Cache` class, further improvements are still needed and will be pulled in a follow-up PR.

Note: code coverage for these functions _may_ go down a little due to the multiple exit points in the functions. I'm not concerned about that as that - again - will be addressed in the follow-up PR.

Includes a dedicated test to verify the cache is used and working.