GH Actions: add new workflow to automate repeated task
The `GetVersionTest` test class contains a constant which should be set to the PHPCS version number of the latest release.

Until now, this version number would be updated manually on an ad-hoc basis.

This workflow automates these updates and will automatically create a PR to update the version number whenever PHPCS has had a new release.

The workflow will run:
* Every night via a cronjob.
    Note: this does mean the workflow will be automatically disabled if there would be no activity in the repo for more than two months.
* On pull requests which update the workflow.
* And can be manually triggered if needed.

Includes removing the update reminder from the release checklist.

See 259, 286, 324.