BackCompat\BCTokens: remove support for PHPCS < 3.7.1
This commit removes support for PHPCS < 3.7.1 for the functionality in the `PHPCSUtils\BackCompat\BCTokens` class.

Note: all previously supported methods still "exist" and will remain functioning and continue to be supported due to the use of the magic `__callStatic()` method.
This is safeguarded via the (updated) tests in the `UnchangedTokenArraysTest` class.

This applies to the following methods:
* `PHPCSUtils\BackCompat\BCTokens::assignmentTokens()`
* `PHPCSUtils\BackCompat\BCTokens::comparisonTokens()`
* `PHPCSUtils\BackCompat\BCTokens::arithmeticTokens()`
* `PHPCSUtils\BackCompat\BCTokens::operators()`
* `PHPCSUtils\BackCompat\BCTokens::parenthesisOpeners()`
* `PHPCSUtils\BackCompat\BCTokens::scopeOpeners()`
* `PHPCSUtils\BackCompat\BCTokens::phpcsCommentTokens()`
* `PHPCSUtils\BackCompat\BCTokens::textStringTokens()`
* `PHPCSUtils\BackCompat\BCTokens::ooScopeTokens()`
* `PHPCSUtils\BackCompat\BCTokens::magicConstants()`