PHP 8.0 | ControlStructures::hasBody(): add support for match expressions
PHP 8.0 introduced match expressions as an new control structure.

While a `match` expression without body or with an empty body is not allowed and would result in a fatal error (or be interpreted as a non-global function call), the `ControlStructures::hasBody()` method should still handle it correctly.

Includes adding the `T_MATCH` token to the `Collections::controlStructureTokens()` token array.

Includes unit tests.

Refs:
* https://wiki.php.net/rfc/match_expression_v2