Utils\MessageHelper::hasNewLineSupport(): remove the method
This method was introduced to support detecting whether new lines in error messages are supported in PHPCS, but is no longer needed now support for PHPCS < 3.7.1 has been dropped.

The method can be safely removed as it was never in a tagged release.

Note: as PHPCS itself does not have any tests for this natively, I'm keeping the associated test and moving it to a separate `Xtra` directory for tests not directly related to PHPCSUtils.