PHP 8.0 | PassedParameters::getParameters(): handle attributes
As of PHP 8.0, attributes can be attached to functions, closures, arrow functions, classes, interfaces, traits, class constants, properties, methods and method parameters.

Attributes effectively instantiate a new instance of the class being referenced and can pass arguments to the constructor of that class.

This has two implications for the `PassedParameters::getParameters()` method:
1. The class instantiation function call within an attribute should be handled by the method same as other class instantiation calls.
2. Attributes may, in select circumstances, be encountered _within_ "parameters" and for efficiency and to prevent incorrect parsing, we should skip over them.

This commit adds a test confirming that situation [1] is already handled correctly.

Additionally, this commit adds the code to skip over attributes _within_ parameters and adds a test covering that as well.

Refs:
* https://wiki.php.net/rfc/attributes_v2